Extracted SudokuFieldsRemover class
